Disable/enable touchpad based on Emacs focus or not
This fixes an issue I have using my new Thinkpad where the touchpad
moves the cursor in Emacs when I'm typing due to pressure from my
hands sitting on the pad. Disabling the touchpad in Emacs only, fixes
this issue and allows me to use for the rest of the system, such as in
a browser where it's very natural to use.